requirements.txt
google-generativeai
streamlit
pandas
jupyter
python-dotenv
pyarrow